Name will now not be overwritten on LogIn
